Se creó un menú de inputs para elegir que realizar + correción de directorios
