--- entregas/Reto001/Intake.java
@@ -1,60 +1,0 @@
-import java.util.Scanner;
-
-class Intake {
-    private FoodNode first;
-
-    public Intake() {
-        first = null;
-    }
-
-    public void addFood(Food food) {
-        FoodNode newFoodNode = new FoodNode(food);
-        if (first == null) {
-            first = newFoodNode;
-        } else {
-            FoodNode current = first;
-            while (current.getNext() != null) {
-                current = current.getNext();
-            }
-            current.setNext(newFoodNode);
-        }
-    }
-
-    public void printFoodListing() {
-        System.out.println(this.toString());
-    }
-
-    public void createIntake() {
-        boolean creating = true;
-        Scanner userInput = new Scanner(System.in);
-        while (creating) {
-            System.out.println("Nombre del alimento (-1 para terminar)");
-            String foodName = userInput.nextLine();
-            if (foodName.equals("-1")) {
-                creating = !creating;
-            } else {
-                Food food = new Food(foodName);
-                addFood(food);
-            }
-        }
-    }
-
-    @Override
-    public String toString() {
-        String foodListing = "";
-        FoodNode current = first;
-        while (current != null) {
-            foodListing = foodListing + current.getFood().toString() + "\n";
-            current = current.getNext();
-        }
-        return foodListing;
-    }
-
-    public static void main(String[] args) {
-
-        Intake breakfast = new Intake();
-        breakfast.createIntake();
-        breakfast.printFoodListing();
-
-    }
-}
